move BrowseApp into own directory
