+++ frontend/src/metabase/browse/containers/BrowseApp.jsx
